Introduced an intermediary service for replacements to fix a problem with circular dependency between StringReplacementsService and AccountsService.
